ledger-on-memory: Move committer construction into the writer.
Closer to where it's used.

CHANGELOG_BEGIN
CHANGELOG_END
